Eliminate boxing warnings
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1232355 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 588e566e093f25530cd677ed4cb87c91f1f64d1a

+++ src/components/org/apache/jmeter/visualizers/StatGraphVisualizer.java
@@ -385,13 +385,13 @@ public class StatGraphVisualizer extends AbstractVisualizer implements Clearable
         graphPanel.setOutlinesBarFlag(drawOutlinesBar.isSelected());
         graphPanel.setShowGrouping(numberShowGrouping.isSelected());
         graphPanel.setLegendPlacement(StatGraphProperties.getPlacementNameMap()
-                .get(legendPlacementList.getSelectedItem()));
+                .get(legendPlacementList.getSelectedItem()).intValue());
 
         graphPanel.setTitleFont(new Font(StatGraphProperties.getFontNameMap().get(titleFontNameList.getSelectedItem()),
-                StatGraphProperties.getFontStyleMap().get(titleFontStyleList.getSelectedItem()),
+                StatGraphProperties.getFontStyleMap().get(titleFontStyleList.getSelectedItem()).intValue(),
                 Integer.parseInt((String) titleFontSizeList.getSelectedItem())));
         graphPanel.setLegendFont(new Font(StatGraphProperties.getFontNameMap().get(fontNameList.getSelectedItem()),
-                StatGraphProperties.getFontStyleMap().get(fontStyleList.getSelectedItem()),
+                StatGraphProperties.getFontStyleMap().get(fontStyleList.getSelectedItem()).intValue(),
                 Integer.parseInt((String) fontSizeList.getSelectedItem())));
 
         graphPanel.setHeight(height);
